Fix editing cabin

+++ src/services/apiCabins.js
@@ -13,36 +13,43 @@ export async function getCabins() {
 export async function createEditCabin(newCabin, id) {
   const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-  const imageName = hasImagePath
+  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+    "/",
+    "",
+  );
+  const imagePath = hasImagePath
     ? newCabin.image
-    : `${Math.random()}-${newCabin.image.name}`.replaceAll("/", "");
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
-
-  // Create cabin
+  // 1. Create/edit cabin
   let query = supabase.from("cabins");
 
+  // A) CREATE
   if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
+  // B) EDIT
   if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
 
   const { data, error } = await query.select().single();
 
   if (error) {
-    console.log(error.message);
-    throw new Error("Cabins could not be created");
+    console.error(error);
+    throw new Error("Cabin could not be created");
   }
 
-  // Upload image
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
-    console.log(storageError.message);
+    console.error(storageError);
     throw new Error(
-      "Cabins image could not be uploaded and the cabin was not created",
+      "Cabin image could not be uploaded and the cabin was not created",
     );
   }
